refactor: move Mask one level up
Also removed useless files
