+++ src/utils/boolToNumber.ts
@@ -6,7 +6,7 @@ import { BitValue } from '../Mask';
  * @param value - The bit to convert.
  * @returns The bit value as a number.
  */
-export function convertToNumber(value: BitValue): number {
+export function boolToNumber(value: BitValue): number {
   let result;
   if (typeof value === 'boolean') {
     result = value ? 1 : 0;
